solve problem of price and decimal

+++ db/migrate/20201201120400_create_carts.rb
@@ -2,7 +2,7 @@ class CreateCarts < ActiveRecord::Migration[5.2]
   def change
     create_table      :carts do |t|
       t.references    :user, null: false, foreign_key: true
-      t.decimal       :stripe_price
+      t.integer       :stripe_price
       t.timestamps
     end
   end
